+++ assertk/src/nativeTest/kotlin/test/assertk/NativeAssertLambdaTest.kt
@@ -2,14 +2,13 @@ package test.assertk
 
 import assertk.assertThat
 import assertk.assertions.*
-import assertk.catch
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.runBlocking
 import kotlin.test.Test
 
 class NativeAssertLambdaTest {
 
-    @UseExperimental(ExperimentalCoroutinesApi::class)
+    @OptIn(ExperimentalCoroutinesApi::class)
     @Test fun returnedValue_works_in_coroutine_test() {
         runBlocking {
             assertThat {
@@ -18,7 +17,7 @@ class NativeAssertLambdaTest {
         }
     }
 
-    @UseExperimental(ExperimentalCoroutinesApi::class)
+    @OptIn(ExperimentalCoroutinesApi::class)
     @Test fun returnedValue_exception_works_in_coroutine_test() {
         runBlocking {
             assertThat {